my-python-app/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.3.0
cffi==1.17.1
discord.py==2.5.2
frozenlist==1.7.0
idna==3.10
multidict==6.5.0
propcache==0.3.2
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.1.0
typing_extensions==4.14.0
yarl==1.20.1
yt-dlp==2025.6.9